Enums, Owl String Fix, Pattern Semantics. Also updated the namespace to the snomed namespace for owl transform.
