+++ plugin/rxnorm-transformation-maven-plugin/src/main/java/dev/ikm/maven/RxnormTransformationMojo.java
@@ -42,7 +42,6 @@ import static dev.ikm.tinkar.terms.TinkarTerm.ENGLISH_LANGUAGE;
 import static dev.ikm.tinkar.terms.TinkarTerm.IDENTIFIER_PATTERN;
 import static dev.ikm.tinkar.terms.TinkarTerm.DEVELOPMENT_PATH;
 import static dev.ikm.tinkar.terms.TinkarTerm.FULLY_QUALIFIED_NAME_DESCRIPTION_TYPE;
-import static dev.ikm.tinkar.terms.TinkarTerm.IDENTIFIER_PATTERN;
 import static dev.ikm.tinkar.terms.TinkarTerm.PREFERRED;
 import static dev.ikm.tinkar.terms.TinkarTerm.REGULAR_NAME_DESCRIPTION_TYPE;
 
@@ -486,343 +485,4 @@ public class RxnormTransformationMojo extends AbstractMojo {
         PrimitiveData.start();
     }
 
-    /**
-     * Extracts RxNorm attributes from the OWL content
-     */
-    private List<RxnormData> extractRxnormData(String owlContent) {
-        List<RxnormData> attributes = new ArrayList<>();
-
-        // Split the content by "# Class:" to identify each concept section
-        String[] classBlocks = owlContent.split("# Class: ");
-
-        // Skip the first block as it's before the first "# Class:"
-        for (int i = 1; i < classBlocks.length; i++) {
-            String block = classBlocks[i];
-            // Extract the URI from the first line
-            // Format: <http://mor.nlm.nih.gov/RXNORM/996062> (meclizine hydrochloride 25 MG Oral Film)
-            Matcher uriMatcher = Pattern.compile("<([^>]+)>").matcher(block);
-
-            if (uriMatcher.find()) {
-                String uri = uriMatcher.group(1);
-                RxnormData concept = new RxnormData(uri);
-
-                // Extract annotations
-                extractAnnotations(block, concept);
-
-                // Extract EquivalentClasses
-                extractEquivalentClasses(block, concept);
-
-                attributes.add(concept);
-            }
-        }
-
-        return attributes;
-    }
-
-    /**
-     * Extracts annotations from a class block
-     */
-    private void extractAnnotations(String block, RxnormData data) {
-        // Extract RxNorm_Name
-        Matcher rxnormNameMatcher = Pattern.compile(":RxNorm_Name <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (rxnormNameMatcher.find()) {
-            data.setRxnormName(rxnormNameMatcher.group(1));
-        }
-
-        // Extract RxNorm_Synonym
-        Matcher rxnormSynonymMatcher = Pattern.compile(":RxNorm_Synonym <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (rxnormSynonymMatcher.find()) {
-            data.setRxnormSynonym(rxnormSynonymMatcher.group(1));
-        }
-
-        // Extract Prescribable_Synonym
-        Matcher prescribableSynonymMatcher = Pattern.compile(":Prescribable_Synonym <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (prescribableSynonymMatcher.find()) {
-            data.setPrescribableSynonym(prescribableSynonymMatcher.group(1));
-        }
-
-        // Extract SNOMED CT identifier
-        Matcher snomedCtMatcher = Pattern.compile("oboInOwl:hasDbXref <[^>]+> \"SNOMEDCT:\\s*([^\"]*)\"").matcher(block);
-        if (snomedCtMatcher.find()) {
-            data.setSnomedCtId(snomedCtMatcher.group(1));
-        }
-
-        // Extract RxCUI identifier
-        Matcher rxCuiMatcher = Pattern.compile("oboInOwl:hasDbXref <[^>]+> \"RxCUI:\\s*([^\"]*)\"").matcher(block);
-        if (rxCuiMatcher.find()) {
-            data.setRxCuiId(rxCuiMatcher.group(1));
-        }
-
-        // Extract VUID identifier
-        Matcher vuidMatcher = Pattern.compile("oboInOwl:hasDbXref <[^>]+> \"VUID:\\s*([^\"]*)\"").matcher(block);
-        if (vuidMatcher.find()) {
-            data.setVuidId(vuidMatcher.group(1));
-        }
-
-        // Extract NDC codes with their start and end dates
-        Matcher ndcMatcher = Pattern.compile("AnnotationAssertion\\(Annotation\\(:endDate \"(\\d+)\"\\) Annotation\\(:startDate \"\\d+\"\\) :ndc <[^>]+> \"([^\"]*)\"\\)").matcher(block);
-        while (ndcMatcher.find()) {
-            String endDate = ndcMatcher.group(1);
-            String ndcCode = ndcMatcher.group(2);
-            data.addNdcCodeWithEndDate(ndcCode, endDate);
-        }
-
-        // Extract QUALITATIVE_DISTINCTION
-        Matcher qualitativeDistinctionMatcher = Pattern.compile(":QUALITATIVE_DISTINCTION <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (qualitativeDistinctionMatcher.find()) {
-            data.setQualitativeDistinction(qualitativeDistinctionMatcher.group(1));
-        }
-
-        // Extract QUANTITY
-        Matcher quantityMatcher = Pattern.compile(":QUANTITY <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (quantityMatcher.find()) {
-            data.setQuantity(quantityMatcher.group(1));
-        }
-
-        // Extract SCHEDULE
-        Matcher scheduleMatcher = Pattern.compile(":SCHEDULE <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (scheduleMatcher.find()) {
-            data.setSchedule(scheduleMatcher.group(1));
-        }
-
-        // Extract HUMAN_DRUG
-        Matcher humanDrugMatcher = Pattern.compile(":HUMAN_DRUG <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (humanDrugMatcher.find()) {
-            data.setHumanDrug(humanDrugMatcher.group(1));
-        }
-
-        // Extract VET_DRUG
-        Matcher vetDrugMatcher = Pattern.compile(":VET_DRUG <[^>]+> \"([^\"]*)\"").matcher(block);
-        if (vetDrugMatcher.find()) {
-            data.setVetDrug(vetDrugMatcher.group(1));
-        }
-
-        // Extract TALLMAN_SYNONYM (can have multiple)
-        Matcher tallmanSynonymMatcher = Pattern.compile(":Tallman_Synonym <[^>]+> \"([^\"]*)\"").matcher(block);
-        while (tallmanSynonymMatcher.find()) {
-            data.addTallmanSynonym(tallmanSynonymMatcher.group(1));
-        }
-
-    }
-
-
-    /**
-     * Extracts EquivalentClasses from a class block
-     */
-    private void extractEquivalentClasses(String block, RxnormData concept) {
-        // Extract the entire EquivalentClasses block with nested parentheses
-        int startIndex = block.indexOf("EquivalentClasses(");
-        if (startIndex != -1) {
-            // Find the matching closing parenthesis by counting opening and closing parentheses
-            int openParenCount = 1;
-            int endIndex = startIndex + "EquivalentClasses(".length();
-
-            while (openParenCount > 0 && endIndex < block.length()) {
-                char c = block.charAt(endIndex);
-                if (c == '(') {
-                    openParenCount++;
-                } else if (c == ')') {
-                    openParenCount--;
-                }
-                endIndex++;
-            }
-
-            if (openParenCount == 0) {
-                // Extract the full block including "EquivalentClasses"
-                String fullEquivalentClasses = block.substring(startIndex, endIndex);
-                concept.setEquivalentClassesStr(fullEquivalentClasses);
-            }
-        }
-    }
-
-    private String readFile(File file) throws IOException {
-        StringBuilder content = new StringBuilder();
-        try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                content.append(line).append("\n");
-            }
-        }
-        return content.toString();
-    }
-    /**
-     * Parses the timestamp from the file name
-     * Expected format: something-YYYY-MM-DD-something.owl
-     */
-    private long parseTimeFromFileName(String fileName) {
-        try {
-            // Extract the date portion using regex
-            Pattern pattern = Pattern.compile("(\\d{4}-\\d{2}-\\d{2})");
-            Matcher matcher = pattern.matcher(fileName);
-
-            if (matcher.find()) {
-                String dateStr = matcher.group(1);
-                SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-                return sdf.parse(dateStr).getTime();
-            } else {
-                LOG.warn("Could not extract date from file name: " + fileName);
-                return System.currentTimeMillis(); // Fallback to current time
-            }
-        } catch (ParseException e) {
-            LOG.error("Error parsing date from file name: " + fileName, e);
-            return System.currentTimeMillis(); // Fallback to current time
-        }
-    }
-
-    /**
-     * Helper class to store RxNorm concept data
-     */
-    private static class RxnormData {
-        private String id;
-        private String uri;
-
-        private String rxnormName = "";
-        private String rxnormSynonym = "";
-        private String prescribableSynonym = "";
-
-        private String snomedCtId = "";
-        private String rxCuiId = "";
-        private String vuidId = "";
-        private List<String> ndcCodes = new ArrayList<>();
-        private Map<String, String> ndcCodesWithEndDates = new HashMap<>();
-        private String qualitativeDistinction = "";
-        private String quantity = "";
-        private String schedule = "";
-        private String humanDrug = "";
-        private String vetDrug = "";
-        private List<String> tallmanSynonyms = new ArrayList<>();
-        private String equivalentClassesStr = "";
-
-        public RxnormData(String uri) {
-            this.uri = uri;
-            if (uri.startsWith("http://mor.nlm.nih.gov/RXNORM/")) {
-                this.id = uri.substring("http://mor.nlm.nih.gov/RXNORM/".length());
-            }
-        }
-
-        public String getId() {
-            return id;
-        }
-
-        public void setRxnormName(String rxnormName) {
-            this.rxnormName = rxnormName;
-        }
-
-        public void setRxnormSynonym(String rxnormSynonym) {
-            this.rxnormSynonym = rxnormSynonym;
-        }
-
-        public void setPrescribableSynonym(String prescribableSynonym) {
-            this.prescribableSynonym = prescribableSynonym;
-        }
-
-        public void setSnomedCtId(String snomedCtId) {
-            this.snomedCtId = snomedCtId;
-        }
-
-        public void setRxCuiId(String rxCuiId) {
-            this.rxCuiId = rxCuiId;
-        }
-
-        public void setVuidId(String vuidId) {
-            this.vuidId = vuidId;
-        }
-
-        public void addNdcCode(String ndcCode) {
-            this.ndcCodes.add(ndcCode);
-        }
-
-        public void setQualitativeDistinction(String qualitativeDistinction) {
-            this.qualitativeDistinction = qualitativeDistinction;
-        }
-
-        public void setQuantity(String quantity) {
-            this.quantity = quantity;
-        }
-
-        public void setSchedule(String schedule) {
-            this.schedule = schedule;
-        }
-
-        public void setHumanDrug(String humanDrug) {
-            this.humanDrug = humanDrug;
-        }
-
-        public void setVetDrug(String vetDrug) {
-            this.vetDrug = vetDrug;
-        }
-
-        public void addTallmanSynonym(String tallmanSynonym) {
-            this.tallmanSynonyms.add(tallmanSynonym);
-        }
-
-        public void setEquivalentClassesStr(String equivalentClassesStr) {
-            this.equivalentClassesStr = equivalentClassesStr;
-        }
-        public void addNdcCodeWithEndDate(String ndcCode, String endDate) {
-            this.ndcCodes.add(ndcCode); // Keep original list for backward compatibility
-            this.ndcCodesWithEndDates.put(ndcCode, endDate);
-        }
-
-        public Map<String, String> getNdcCodesWithEndDates() {
-            return ndcCodesWithEndDates;
-        }
-
-
-        public String getRxnormName() {
-            return rxnormName;
-        }
-
-        public String getRxnormSynonym() {
-            return rxnormSynonym;
-        }
-
-        public String getPrescribableSynonym() {
-            return prescribableSynonym;
-        }
-
-        public String getSnomedCtId() {
-            return snomedCtId;
-        }
-
-        public String getRxCuiId() {
-            return rxCuiId;
-        }
-
-        public String getVuidId() {
-            return vuidId;
-        }
-
-        public List<String> getNdcCodes() {
-            return ndcCodes;
-        }
-
-        public String getQualitativeDistinction(){
-            return qualitativeDistinction;
-        }
-
-        public String getQuantity(){
-            return quantity;
-        }
-
-        public String getSchedule(){
-            return schedule;
-        }
-
-        public String getHumanDrug() {
-            return  humanDrug;
-        }
-
-        public String getVetDrug() {
-            return vetDrug;
-        }
-
-        public List<String> getTallmanSynonyms(){
-            return tallmanSynonyms;
-        }
-
-        public String getEquivalentClassesStr() {
-            return equivalentClassesStr;
-        }
-    }
 }
